Public-portfolio/Finance/Portfolio Optimisation/requirements.txt
yfinance
matplotlib
pandas
numpy
scipy
fredapi
python-docx
python-dotenv
